+++ NEWS.md
@@ -1,3 +1,5 @@
+# vroom (development version)
+
 # vroom 1.3.0
 
 * The Rcpp dependency has been removed in favor of cpp11.